rebase to refactor/gmc_dataprocessing

+++ multigeiger/webconf.h
@@ -23,7 +23,7 @@ extern char appkey[];
 extern float localAlarmThreshold;
 extern int localAlarmFactor;
 
-extern int sendDataToMessengerEvery;
+extern long sendDataToMessengerEvery;
 extern char telegramBotToken[50];
 extern char telegramChatId[15];
 extern bool sendLocalAlarmToMessenger;
